refactor(event): moved RabbitMQEventConsumer from adapters to entrypoints
- Relocate RabbitMQEventConsumer class from adapters/eventconsumer.py to entrypoints/eventconsumer.py
- Remove unused adapters/eventconsumer.py module
- Update imports in affected files

This change resolves:
1. Misclassification of RabbitMQEventConsumer as an adapter
2. Lack of consumption logic in entrypoints layer
